Fix one more review note.
